Switch libfaac to single-precision float arithmetic
This commit replaces the use of 'double' with 'float' throughout the libfaac library. Standard math functions have been updated to their single-precision counterparts (e.g., sinf, cosf, sqrtf). This optimization targets platforms with limited double-precision performance, such as MIPS Xburst1, without introducing acoustic regressions. Bit-identical output has been verified for several test signals.

+++ libfaac/coder.h
@@ -77,8 +77,8 @@ typedef struct {
     int direction;                       /* Filtering direction */
     int coefCompress;                    /* Are coeffs compressed? */
     int length;                          /* Length, in bands */
-    double aCoeffs[TNS_MAX_ORDER+1];     /* AR Coefficients */
-    double kCoeffs[TNS_MAX_ORDER+1];     /* Reflection Coefficients */
+    float aCoeffs[TNS_MAX_ORDER+1];     /* AR Coefficients */
+    float kCoeffs[TNS_MAX_ORDER+1];     /* Reflection Coefficients */
     int index[TNS_MAX_ORDER+1];          /* Coefficient indices */
 } TnsFilterData;
